requirements.txt
fastapi
uvicorn
python-dotenv
numpy
torch
transformers